+++ src/pymyhondaplus/__init__.py
@@ -3,14 +3,17 @@
 from .api import AuthTokens, CommandResult, EVStatus, Geofence, HondaAPI, HondaAPIError, HondaAuthError, SubscriptionService, Subscription, UIConfiguration, UserProfile, Vehicle, VehicleCapabilities, compute_trip_stats, parse_charge_schedule, parse_climate_schedule, parse_ev_status
 from .auth import DeviceKey, HondaAuth, encrypt_request
 from .storage import SecretStorage, get_storage
+from .translations import CAPABILITY_API_KEY_TO_TRANSLATION_KEY, TRANSLATIONS, get_translator
 
 __all__ = [
     "AuthTokens",
+    "CAPABILITY_API_KEY_TO_TRANSLATION_KEY",
     "CommandResult",
     "EVStatus",
     "Geofence",
     "Subscription",
     "SubscriptionService",
+    "TRANSLATIONS",
     "UIConfiguration",
     "UserProfile",
     "Vehicle",
@@ -19,6 +22,7 @@ __all__ = [
     "HondaAPIError",
     "HondaAuthError",
     "compute_trip_stats",
+    "get_translator",
     "parse_charge_schedule",
     "parse_climate_schedule",
     "parse_ev_status",